requirements.txt
streamlit
groq
pypdf
python-dotenv
fpdf